Move ScheduleNote file to added courses
